Split extensions by dedicated files
