+++ api/src/main/kotlin/io/spine/protodata/PrimitiveTypeExts.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022, TeamDev. All rights reserved.
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,39 +24,13 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.protodata.java.file
+@file:JvmName("PrimitiveTypes")
 
-import io.spine.protodata.Option
-import io.spine.protodata.TypeInstances.boolean
-import io.spine.protodata.TypeInstances.string
-import io.spine.protodata.option
-import io.spine.protodata.pack
-import io.spine.protodata.packedTrue
+package io.spine.protodata
 
 /**
- * The option to instruct `protoc` to generate multiple Java files.
+ * Obtains a [Type] wrapping this `PrimitiveType`.
  */
-public val javaMultipleFiles: Option = option {
-    name = "java_multiple_files"
-    type = boolean
-    value = packedTrue
-}
-
-/**
- * Obtains an option to set the Java package with the given [name]
- * for the generated code.
- */
-public fun javaPackage(name: String): Option = option {
-    this.name = "java_package"
-    type = string
-    value = name.pack()
-}
-
-/**
- * Obtains the option to set the [name] of the outer Java class.
- */
-public fun javaOuterClassName(name: String): Option = option {
-    this.name = "java_outer_classname"
-    type = string
-    value = name.pack()
+public fun PrimitiveType.asType(): Type = type {
+    primitive = this@asType
 }
